Refactorized, added folders and inside index.js and style.css
