+++ EventManager_Frontend/event.manager.client/src/App.js
@@ -1,9 +1,8 @@
 import React from 'react';
-import {render} from 'react-dom';
-import {Label, FormControl, Button, ButtonToolbar} from 'react-bootstrap';
-import HomeNavbar from './components/Navbar';
-import Home from './components/Home';
-import NewEvent from './components/NewEvent';
+import HomeNavbar from './components/Navbar/index';
+import NewLecture from './components/NewLecture/index';
+import Home from './components/Home/index';
+import NewEvent from './components/NewEvent/index';
 import {BrowserRouter as Router,Route, Switch } from 'react-router-dom';
 
 class App extends React.Component {
@@ -13,11 +12,12 @@ class App extends React.Component {
           <HomeNavbar/>
           <Router>
                   <Switch>
-                      <Route exact path="/" component={Home}/>
+                      <Route path="/" component={Home}/>
+                      <Route exact path="/home" component={Home}/>
+                      <Route exact path="/NewEvent" component={NewEvent}/>
+                      <Route exact path="/NewLecture" component={NewLecture}/>
                   </Switch>
           </Router>
-
-
       </div>
     );
   }
