+++ EventManager_Frontend/event.manager.client/src/components/NewLecture/index.js
